clean zipcodes and type cast

+++ facdb/sql/_deduplication.sql
@@ -21,7 +21,7 @@ AND bin in (
 );
 
 -- Remove out of NYC zipcodes
-DELETE FROM facdb WHERE zipcode NOT IN (SELECT zipcode FROM doitt_zipcodeboundaries);
+DELETE FROM facdb WHERE zipcode NOT IN (SELECT zipcode::integer FROM doitt_zipcodeboundaries);
 
 -- Remove records outside of NYC based on geometry
 DELETE FROM facdb WHERE geom IS NOT NULL AND uid NOT IN (
